+++ patches/src/main/kotlin/app/revanced/patches/duolingo/music/Fingerprints.kt
@@ -1,0 +1,8 @@
+package app.revanced.patches.duolingo.music
+
+import app.revanced.patcher.fingerprint
+
+// Matches toString()
+internal val musicCourseFingerprint = fingerprint {
+    strings("Music(courseSummary=", ", activePathSectionId=")
+}