+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# [1.2.0-dev.3](https://github.com/hoo-dles/revanced-custom-patches/compare/v1.2.0-dev.2...v1.2.0-dev.3) (2025-06-05)
+
+
+### Features
+
+* **Merriam-Webster:** Add `Enable Premium` patch ([92adc45](https://github.com/hoo-dles/revanced-custom-patches/commit/92adc45b4cb0c97886f914bb8c14b04529882541))
+
+# [1.2.0-dev.2](https://github.com/hoo-dles/revanced-custom-patches/compare/v1.2.0-dev.1...v1.2.0-dev.2) (2025-06-04)
+
+
+### Features
+
+* **Duolingo:** Add `Unlock licensed songs` patch ([e45b8c1](https://github.com/hoo-dles/revanced-custom-patches/commit/e45b8c1b36f707f46e59b3bbc9b91191466011d4))
+
+# [1.2.0-dev.1](https://github.com/hoo-dles/revanced-custom-patches/compare/v1.1.0...v1.2.0-dev.1) (2025-06-03)
+
+
+### Features
+
+* **Duolingo:** Add `Unlock super` patch ([c80241d](https://github.com/hoo-dles/revanced-custom-patches/commit/c80241d91f423aed48a5b1f7cb730c0d426febb1))
+* **Duolingo:** Adding `Disable practice reminder nag` patch ([179ab99](https://github.com/hoo-dles/revanced-custom-patches/commit/179ab992c264d418d4e910105fbef9a300d36ede))
+
 # [1.1.0](https://github.com/hoo-dles/revanced-custom-patches/compare/v1.0.0...v1.1.0) (2025-05-18)